// src/app/kimios-client-api/model/dMEntityTree.ts
/**
 * Kimios API
 * Kimios Sample API
 *
 * OpenAPI spec version: 1.3-SNAPSHOT - d5738c4d578e4ed21beaebdff49ff1159a7d86a5
 * Contact: documentation@kimios.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DMEntityTreeNode } from './dMEntityTreeNode';


export interface DMEntityTree { 
    treeNodeList?: Array<DMEntityTreeNode>;
}